requirements.txt
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-dotenv==1.0.1
tabulate==0.9.0
